Move max send amount methods to UI to resolve package too large issues
