+++ app/scripts/controllers/transactions/bitbox-utils.js
@@ -105,62 +105,6 @@ class BitboxUtils {
     })
   }
 
-  static removeUnspendableUtxo(utxo) {
-    const sorted = utxo.sort((a, b) => {
-      return b.satoshis - a.satoshis
-    })
-
-    const spendable = sorted.map(x => {
-      if (x.spendable) {
-        return x
-      }
-    })
-
-    const clean = spendable.filter(val => {
-      return val !== undefined
-    })
-
-    const chunk = this.chunk(clean, 20)
-
-    // limit to 20 utxo to prevent tx failing
-    return chunk[0]
-  }
-
-  static chunk(arr, size) {
-    return Array.from({ length: Math.ceil(arr.length / size) }, (v, i) =>
-      arr.slice(i * size, i * size + size)
-    )
-  }
-
-  static async calculateMaxSendSatoshis(spendableUtxos) {
-    if (!spendableUtxos || spendableUtxos.length === 0) {
-      throw new Error('Insufficient funds')
-    }
-
-    // Calculate fee
-    let byteCount = 0
-    const sortedSpendableUtxos = spendableUtxos.sort((a, b) => {
-      return b.satoshis - a.satoshis
-    })
-    const inputUtxos = []
-    let totalUtxoAmount = 0
-    for (const utxo of sortedSpendableUtxos) {
-      if (utxo.spendable !== true) {
-        throw new Error('Cannot spend unspendable utxo')
-      }
-      inputUtxos.push(utxo)
-      totalUtxoAmount += utxo.satoshis
-
-      byteCount = SLP.BitcoinCash.getByteCount(
-        { P2PKH: inputUtxos.length },
-        { P2PKH: 2 }
-      )
-    }
-
-    const maxSendSatoshis = totalUtxoAmount - byteCount
-    return maxSendSatoshis
-  }
-
   static signAndPublishBchTransaction(txParams, spendableUtxos) {
     return new Promise(async (resolve, reject) => {
       try {
